+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+- Importing more than 10 subgroups into a realm
+
 ## [6.0.2] - 2024-06-17
 - Restored versioning